factory/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
futures==3.0.5
idna==3.6
jsonrpclib==0.1.7
markdown2==2.4.10
mock-server==0.3.9
py-bcrypt==0.4
python-dotenv==1.0.0
requests==2.31.0
six==1.10.0
tornado==4.3
urllib3==2.1.0
